refactor: move main window to views dir
