+++ src/views/mod.rs
@@ -1,0 +1,4 @@
+mod home;
+mod settings;
+
+pub use home::HomeView;
